refactor(catalog): separate the two preview paths and scope the probe session
Cleanup over the paired preview work. No behaviour is added.

The raster path and the live path had ended up in one module sharing nothing
but a subject. `LibraryPreviewViewport` -- react-zoom-pan-pinch over a stored
SVG, still what the Assets tab thumbnails and the revision visual diff use --
moves to `library-preview-viewport.tsx`, leaving the inspector to the live
canvas pair.

The lightbox now opens its own cross-probe session instead of nesting in the
embedded pair's. One shared session registered all four viewers, so every
pointer move also repainted the two canvases sitting behind the modal, and
latching a pin in the lightbox changed the badge on the page underneath.

`BaseRenderOptions.interactive` was a migration alias for a migration that
finished; nothing passed it, and the renderer bundle drops its half in the
same change.

Also: the preview cards gate on the representation pair they actually draw
rather than on the raw asset list, so a component whose assets are not linked
into a representation no longer advertises renderable assets and then shows two
empty panes; `LibraryAssetRenderer` calls `onUnitsChange` directly instead of
through a `useCallback` that only forwarded it; and the duplicated
`--preview-surface` says why it is duplicated.
